__init__.py
"""
Bug Triage Environment — package init.

Re-exports the public API so users can do:

    from bug_triage_env import BugTriageEnv, TriageAction, TriageObservation
"""

try:
    from bug_triage_env.client import BugTriageEnv
    from bug_triage_env.models import TriageAction, TriageObservation
except ImportError:
    from client import BugTriageEnv
    from models import TriageAction, TriageObservation

__all__ = ["BugTriageEnv", "TriageAction", "TriageObservation"]
